feat: implement main feature temporary
All of the input processing and domain logic is now implemented in the
main function. But it's very hard to handle exceptional situation in
this implementation.

It will be better to make a seprate class to handle input processing
that thorws exceptions with proper messages, then all exceptions thrown
by the instances can be handled in the main function in more integrated
manner.