feat($LottoView, $LottoController, $LottoLogic, $Lotto, $ValidationForm, $LottoEnum) : implement 기능2
LottoView gets inputs, winning numbers and bonus number. It checks not-a-business-logic exceptions, like number format, count, etc.

LottoModel has other methods to validate inputs, validateWinningNumbers and validateBonusNumber. It's because we need to ensure reenter the value in each input process.

ValidationForm is a class to confirm that a value is validated with other method.
LottoEnum is a enum to rank how many numbers of a lottery is matched with winning numbers.